fix angle_axis_proper (...), add test
