+++ src/third_party_addendum.jl
@@ -1,23 +1,29 @@
-rotate(x::SMatrix{3}, q::Quaternion) = rotationmatrix_normalized_fsa(q) * x
+function vector_to_skew_symmetric{T}(v::SVector{3, T})
+    @SMatrix [zero(T) -v[3] v[2];
+              v[3] zero(T) -v[1];
+              -v[2] v[1] zero(T)]
+end
+
+const hat = vector_to_skew_symmetric
+
+function cross(a::SVector{3}, B::AbstractMatrix)
+    vector_to_skew_symmetric(a) * B
+end
+
+rotate(x::SMatrix{3}, q::Quaternion) = rotation_matrix(q) * x
 
 function rotate(x::SVector{3}, q::Quaternion)
     qret = q * Quaternion(zero(eltype(x)), x[1], x[2], x[3]) * inv(q)
     SVector(qret.v1, qret.v2, qret.v3)
 end
 
-angle_axis_proper(q::Quaternion) = angle_proper(q), axis_proper(q)
-
-angle_proper(q::Quaternion) = 2 * acos(real(Quaternions.normalize(q)))
-
-function axis_proper(q::Quaternion)
-    q = Quaternions.normalize(q)
-    s = sin(angle(q) / 2)
-        abs(s) > 0 ?
-        [q.v1, q.v2, q.v3] / s :
-        [1.0, 0.0, 0.0]
+function angle_axis_proper{T}(q::Quaternion{T})
+    Θ = 2 * atan2(√(q.v1^2 + q.v2^2 + q.v3^2), q.s)
+    axis = isapprox(Θ, zero(T)) ? SVector{3, T}(1, 0, 0) : SVector{3, T}(q.v1, q.v2, q.v3) / sin(Θ / 2)
+    Θ, axis
 end
 
-function rotationmatrix_normalized_fsa{T}(q::Quaternion{T})
+function rotation_matrix{T}(q::Quaternion{T})
     sx, sy, sz = 2q.s*q.v1, 2q.s*q.v2, 2q.s*q.v3
     xx, xy, xz = 2q.v1^2, 2q.v1*q.v2, 2q.v1*q.v3
     yy, yz, zz = 2q.v2^2, 2q.v2*q.v3, 2q.v3^2
